preserve calibration data during rauc updates
Only the bbb-based calibration-box currently uses these. I think we'll
move stuff to EEPROMs, eventually, but before that happens, let's make
sure the data are not lost upon update.

I forgot to add this one because the first bring-up used just that SD
card, so there were no "updates" to speak of.

Change-Id: I669547cd88d25779e59486c702782a62afa79756