calibration-box: allow for longer HW timeouts
This HW is "magic", a write can take possibly very long time to
complete, Increase the systemd service watchdog interval.

This does not solve a problem of a sysrepo call timing out, but at least
it won't take the whole service with it (that would actually block *any*
progress, because with this patch we will eventually switch that damned
laser on).

Change-Id: I4f7efc7bbf47ba6d2ad6b98ad209f4fdfb6cde09