Shave off about a minute from builds
BR2_REPRODUCIBLE performs several things related to timestamps,
hostnames, etc. One extra thing which it does is disabling
multithreading in `xz` compression of the rootfs tarball. On my laptop
with a i5-6300U CPU and an almost no-op rebuild (`time make
cla-sysrepo-reconfigure rootfs-czechlight-rauc`), this item has just
shortened the build time from 3m10s to 2m13s.

A CI build prior to this patch takes a little less than 20min. If the
speedup is at least as good, it means that the build will become at
least 5% faster, and that's IMHO a pretty nice feature.

Change-Id: I8bb22f388c17bbb9ef56e989cd7255ec4cee1359