CI: clearfog: keep the work dir as an artifact
A regular build (i.e., a change uploaded to br2-external) should also
create a compressed tarball of Buildroot's working directory. This will
be used later on as a starting state for jobs triggered by, say,
cla-sysrepo.

Interestingly, I'm not getting a `staging/` directory when in CI, but
I'm seeing it locally. The docs say it's a backward-compatible thingy,
so let's just skip it.

Also, do not upload the U-Boot SPL when being triggered externally.

Change-Id: I7ac4edf6d1775192e17f0cefc77c25dac979b9e6