Register this project for *upstream* consumers
This is important to do before actually bringing in
CzechLight/dependencies via either cla-sysrepo *or* netconf-cli. That
way changes in these two projects can depend on this change which makes
that repo available, and then the next change in this repo can depend on
changes in these leaf projects. Sweet, isn't it.

Change-Id: I0123f48551cbcf5cb2f90f86e1c7f8bc7c80051b