new sysrepo
Updating all dependencies so that we're using the new sysrepo. The easiest way
there was via bringing in the new buildroot as well, which meant:

- new RAUC, hence no patch needed anymore,
- newer systemd, hence an updated patch,
- New spdlog and fmt, which means that the version we ship in
  submodules/dependencies no longer works. Fix that by just relying on
  systemwide spdlog.

FIXME: update dependencies so that spdlog is not bundled anymore

FIXME: it doesn't boot. Bummer.

FIXME: we'll have to update the kernel as well, but that's currently broken.
This is kinda important due to Fedora 33 and BR2_PACKAGE_LINUX_TOOLS_PERF=y
because that version won't build with GCC 10.2 on Fedora 33.

Change-Id: I4b65a8fb345bfe7907a331d0de16294af1c36a78
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/dependencies/+/2884
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/3010
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/netconf-cli/+/2885
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/gammarus/+/3122
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/lldp-systemd-networkd-sysrepo/+/3118