clearfog: update kernel to 5.9.1
This was a funny one:

- On 5.9, the PCIe interface on Clearfog does not work at all due to
some PCIe range overlap with something else [1].

- The patch [2] which converted spi-orion.c to use `gpiod` instead of
that plain-old GPIO was apparently broken, *OR* perhaps it is not that
patch, but just a result of the general GPIO brokeness on mvebu that I
reported earlier [3]. Anyway, I'm simply reverting that.

- There were also some mcp23s08-related segfaults introduced in 5.9.

Also, there's a bunch of other patches, and as always, these are
available as a git tree [4], too.

[1] https://patchwork.ozlabs.org/project/linux-pci/patch/20201023145252.2691779-1-robh@kernel.org/
[2] https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux/+/66eb228988cf82abb5f633a7d585120923308caf
[3] https://www.spinics.net/lists/linux-gpio/msg49678.html
[4] https://gerrit.cesnet.cz/plugins/gitiles/github/torvalds/linux/+log/refs/heads/cesnet/2020-10-26---5.9.1

Change-Id: I8b229cb0e86fdbfc5358452054354cf8a786c961