Use new, daemon-less sysrepo for YANG management
Updating all dependencies so that we're using the new sysrepo. The easiest way
there was via bringing in the new buildroot as well, which meant:

- new RAUC, hence no patch needed anymore,
- newer systemd, hence an updated patch,
- New spdlog and fmt, which means that the version we ship in
  submodules/dependencies no longer works. Fix that by just relying on
  systemwide spdlog.
- uboot-tools needed a patch because there's no /var/lock,
- OpenSSH was previously started autmagically, now we're explicitly
  adding that to the list of packages (I really like SSH, don't you?).

New sysrepo means that there's no `sysrepod` anymore, which required
changing a bunch of other units so that they are not marked as
`PartOf=sysrepod.service` anymore. Also, the command-line options for
`sysrepocfg` and `sysrepoctl` changed.

I had to patch buildroot so that it doesn't create SSH keys for the
NETCONF server at build time, and that all required modules are actually
installed. It turns out that new sysrepo doesn't really install
anything, and the fun only starts with Netopeer2. Their install scripts
a wee bit picky, but hey, we have bash, we can make it work.

On the other hand, installation of *our* YANG modules is a bit hairy.
There's just too much parallel activities going on, and therefore I have
no idea how robust our callout to --apply really is. As a cherry on top,
`sysrepocfg` is particularly inconsistent when it comes to how the ****
one is supposed to provision the initial config, etc etc. In the end, I
ended up with a big hammer.

FIXME: Install all non-netopeer2 YANG modules from just a single script

FIXME: we'll have to update the kernel as well, but that's currently broken.
This is kinda important due to Fedora 33 and BR2_PACKAGE_LINUX_TOOLS_PERF=y
because that version won't build with GCC 10.2 on Fedora 33.

Change-Id: I4b65a8fb345bfe7907a331d0de16294af1c36a78
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/3130
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/netconf-cli/+/3135
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/lldp-systemd-networkd-sysrepo/+/3136
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/velia/+/3137
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/gammarus/+/3122