shell: add a nice prompt
I'm doing this especially for the hostanme, because it's super-handy to
know what box I'm on, especially when hacking around in the lab with
USB-serial jumping where the SSH (and therefore xterm title) doesn't
necessarily matches the final target node.

Yellow is the color that we used to have in that new shiny logo before
it became grayscale, and I kinda like that, so let's make it yellow.

Change-Id: I9a64da775f330d4a357e84ba2320f30f52c7dd71