Do not indicate failures when log upload fails
We have this velia thingy which lights up a red LED when something fails
within the system. It turns out that some failures are, however, not
really a *fault* of the system. For example, when a central log server
gets rebooted and the system is configured to upload debugging logs,
there's no point in treating that as a "system failure".

Fix this by selectively blacklisting some systemd services and ignoring
their transient failures (or even permanent ones, really, we have no way
of distinguishing among them anyway).

Change-Id: I0649c61454b99a7912f88307bf5e0273f08c0067