docs: gdb: do not use incomplete sysroot
That path was from the ARM (or Linaro?) toolchains. We've switched to
the Bootlin ones, and they do not ship this file. There's another one,
`host/arm-buildroot-linux-gnueabihf/sysroot/usr/share/buildroot/gdbinit`,
but it only sets sysroot, and the sysroot it uses only contains the
toolchain's own sysroot. The full command already sets a sysroot with
the `target/` directory, which is sufficient, as it also gets whatever
the toolchain distributes.

Change-Id: I92d590b14e955d70398a8f7af8e9ba3fc98c43bb