clearfog: do not support I2C address translation via LTC4316
It was a nice idea, but the power supplies we're using insist on using
that extra checksum byte (PEC) which also includes the I2C address.
There's nothing in the kernel today which can rewrite the PEC for a new
checksum, and while that would not be terribly hard to do, it just is
not there yet. Also, the rev2 of these boards does not have LTC4316 at
all, and we haven't been using that chip for the past year or so (it's
always been muxed to that bit-banging sofwtare I2C).

Change-Id: I8073f30501050b4a148fead87e0b1ece24130a04