test: Update cla-sysrepo and velia
We have split ietf-hardware-state implementation between two daemons,
cla-sysrepo and velia. This commit makes br2-external to use the
versions of cla-sysrepo and velia after these changes.

Velia also now uses an appliance parameter with similar use to
cla-sysrepo.  Only that (as of now) we share a common appliance model
for every czechlight-clearfog device.
Velia now recognizes it is running on czechlight-clearfog and initializes
the sysfs drivers correctly for this board.

Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/cla-sysrepo/+/3326
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/3326
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/velia/+/3323
Depends-on: https://cesnet-gerrit-public/c/CzechLight/velia/+/3323
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/velia/+/3323
Change-Id: I783adacfe23f2d00c6398f3e7a598ba58babeaad