yang: install ietf-hardware instead of ietf-hardware-state
With latest changes to cla-sysrepo and velia, we should install
ietf-hardware model instead of ietf-hardware-state now.

Unfortunately, we encountered a bug/feature of Sysrepo. When
uninstalling implemented model which has an imported dependency, then
the YANG model of the imported dependency is not removed from the
repository. See example below.

We have two implemented models. iana-hardware and ietf-hardware-state.
ietf-hardware model is imported as a dependency of ietf-hardware-state.

 # pwd is sysrepo repository

 $ sysrepoctl -l | grep hardware | cut -d"|" -f1-3
 iana-hardware              | 2018-03-13 | I
 ietf-hardware              | 2018-03-13 | i
 ietf-hardware-state        | 2018-03-13 | I

 $ ls *hardware*
 iana-hardware@2018-03-13.yang  ietf-hardware-state@2018-03-13.yang  ietf-hardware@2018-03-13.yang

 $ sysrepoctl -u ietf-hardware-state --apply

 $ sysrepoctl -l | grep hardware | cut -d"|" -f1-3
 iana-hardware              | 2018-03-13 | I

 $ ls *hardware*
 iana-hardware@2018-03-13.yang  ietf-hardware@2018-03-13.yang

Change-Id: I580e52ff10c5a17f99db18c080b10c1d40aacd78