Update cla-sysrepo and velia to use ietf-hardware
Update cla-sysrepo and velia. They now implement ietf-hardware YANG
model instead of ietf-hardware-state (both defined in RFC 8348 [1]).

With latest changes to cla-sysrepo and velia, we should install
ietf-hardware model instead of ietf-hardware-state now.

Unfortunately, we encountered a bug/feature of Sysrepo. When
uninstalling implemented model which has an imported dependency, then
the YANG model of the imported dependency is not removed from the
repository. See example below.
Link to Sysrepo's issue tracker is in the footer.

We have two implemented models. iana-hardware and ietf-hardware-state.
ietf-hardware model is imported as a dependency of ietf-hardware-state.

 # pwd is sysrepo repository

 $ sysrepoctl -l | grep hardware | cut -d"|" -f1-3
 iana-hardware              | 2018-03-13 | I
 ietf-hardware              | 2018-03-13 | i
 ietf-hardware-state        | 2018-03-13 | I

 $ ls *hardware*
 iana-hardware@2018-03-13.yang  ietf-hardware-state@2018-03-13.yang  ietf-hardware@2018-03-13.yang

 $ sysrepoctl -u ietf-hardware-state --apply

 $ sysrepoctl -l | grep hardware | cut -d"|" -f1-3
 iana-hardware              | 2018-03-13 | I

 $ ls *hardware*
 iana-hardware@2018-03-13.yang  ietf-hardware@2018-03-13.yang

[1] https://tools.ietf.org/html/rfc8348

Bug: https://github.com/sysrepo/sysrepo/issues/2263
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/cla-sysrepo/+/3352
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/3352
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/velia/+/3329
Depends-on: https://cesnet-gerrit-public/c/CzechLight/velia/+/3329
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/velia/+/3329
Change-Id: I580e52ff10c5a17f99db18c080b10c1d40aacd78