update and change sysrepo persistency
In sysrepo commit c226ab1 (November 2020), some fields in the internal
sysrepo-monitoring module were renamed from pid to cid without a
revision bump of that module. If a post-November sysrepo ever runs on a
system with pre-November YANG module, sysrepo operations fail with this
error:

 Failed to find "cid" as a sibling to "sysrepo-monitoring:datastore".

...and we started hitting just that in our lab.

The root cause appears to be a bug in our YANG installation where we
accidentally preserve not just *.persist files, but the whole of
/etc/sysrepo.

Also sync all packages.

Change-Id: Ib2060ea850d1121d5025183b6c5b17d8962f435b
Depends-on: https://cesnet-gerrit-czechlight/c/CzechLight/gammarus/+/3367
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/gammarus/+/3367
Bug: https://tree.taiga.io/project/jktjkt-netconf-cli/issue/206
Bug: https://github.com/sysrepo/sysrepo/issues/2280