CI: rebuild the NETCONF stack if updated via CzechLight/dependencies
If there's a change for, say, CzechLight/cla-sysrepo which depends-on a
change for CzechLight/dependencies, the build job which rebuilds the FW
image for that change of cla-sysrepo was ignoring updated dependencies
by mistake.

As explained in the commit message, this is far from perfect, but hey,
it's a CI system, not a spaceship.

Change-Id: If590c2dc8733d7960784e171c06a005511ee3277