systemd: RemainAfterExit=yes for all Type=oneshot units
Many of these are intended to run as a one-time thingy. Some of them,
such as the YANG installation bits, have been failing since the recent
switch to storing sysrepo config as a JSON file (and the related change
of not preserving /etc/sysrepo/ directly).

One especially nasty scenario is a random failure in cla-sysrepo which
leads to an automatic restart, which triggers czechlight-install-yang,
which fails horribly. Yay.

Change-Id: I76664c5fb2b8eb62a12fd19a31fc606fa7c3723c