Enable coredumps from systemd units
On ARM, backtraces on C code are not terribly useful by default because
the stack cannot really be unwound, hence a custom compiler flag for
this. This has never been a problem because I've only ever debugged C++
code *on the device itself*, and for C++ code stack unwinding is enabled
by the language frontend because it's needed for exception support. On
the C code, not so much.

With this flag, U-Boot fails to build, but just its examples. There's no
need to help these examples in the first place, though, so let's just
screw^Wskip them.

I *think* I also need to explicitly enable non-zero core sizes, even if
I already have systemd-coredumpctl enabled via systemd magic.

Change-Id: I094c52c2ead9c1dd581de4974e9edd432c8b6894