clearfog: Upgrade U-Boot
Updating U-Boot is not something that we do every day; in fact, I wanted
to *never ever* do that after the initial deployment. However, now it
seems to be a good time because all of the devices are being upgraded
with new EDFA modules and especially the updated electronics, so let's
see if sticking with an U-Boot that's more than two years newer improves
the experience with USB flash drives. Also, I'm really hoping that this
will reduce the number of nasty surprises when we update the compiler
and the dtc next time.

"Unfortunately", it's right after the 2021.01 release, which means that
there are hundreds of commits in master since that tag already. When
we're updating, be sure to update the hell of the system, so let's bump
it all the way to 11. Hopefully, we can catch something like 2021.04
when it gets released.

Change-Id: I8730956f43731e777376f34e795a15e30cbfbf09