clearfog: Upgrade U-Boot
Updating U-Boot is not something that we do every day; in fact, I wanted
to *never ever* do that after the initial deployment. However, now it
seems to be a good time because all of the devices are being upgraded
with new EDFA modules and especially the updated electronics, so let's
see if sticking with an U-Boot that's more than two years newer improves
the experience with USB flash drives (it does). Also, I'm really hoping
that this will reduce the number of nasty surprises when we update the
compiler and the dtc next time.

"Unfortunately", it's right after the 2021.01 release, which means that
there are hundreds of commits in master since that tag already. When
we're updating, be sure to update the hell of the system, so let's bump
it all the way to 11. Hopefully, we can catch something like 2021.04
when it gets released.

The watchdog timer was a bit tricky. Apparently, it's an unsigned 32bit
counter ticking at 25 MHz, giving us about 171s of breathing time until
the watchdog shoots us into the head. In the past I was bitten by too
short timeouts (most recently earlier today), so let's use this long
default. The time can be made shorter later on in Linux, if needed (not
sure if it's possible via a U-Boot command).

Change-Id: I8730956f43731e777376f34e795a15e30cbfbf09