clearfog: U-Boot: export version information
The goal is to make this information available in the userspace. We're
building from git, but Buildroot eats that, and I don't want to get
another repo in there via Zuul and set up hooks for preserving the .git
dir, so let's just hardcode that.

The `ver` variable is magic. Even if it's stored in the uboot env, on
the next boot it get magically overwritten in U-Boot's runtime. Given
that our bootcmd then invokes `saveenv`, the userspace can access that
value just as well, and there's no need for passing it via kernel, or
something like that.

Change-Id: Iddbd232dd725f11f511abd55270443607e22bc95