clearfog: U-Boot: export version information
The goal is to make this information available in the userspace. We're
building from git, but Buildroot eats that, and I don't want to get
another repo in there via Zuul and set up hooks for preserving the .git
dir, so let's just hardcode that. On top of that, `ver` is a transient
thingy, not a part of the boot env, so let's pass it to the booted
kernel via its command line.

Change-Id: Iddbd232dd725f11f511abd55270443607e22bc95