Switch the local CLI to use NETCONF over Unix socket
Without this patch, any local user could well reboot the box via an RPC.
That's not what we want. Because sysrepo doesn't do NACM these days,
but delegates that to the netopeer2-server instead, we have to talk
NETCONF in order to validate everything.

Of course this currently Doesn't Work At All™ because netopeer2-server
segfaults, but hey, that's another problem.

Bug: https://github.com/CESNET/netopeer2/issues/821
Bug: https://github.com/sysrepo/sysrepo/issues/2312
Change-Id: I9c100fd73f11b4b6c759ee382e82c9e75f3cea8f