Update kernel to 5.11.3
This one was funny due to some changes in gpiolib. On the other hand, it
seems that we no longer need the PCIe fixups or the gpio/pinctrl core
fixups, which is awesome.

Patches are based on v5.11.3, and everything is in Gerrit in the
cesnet/2021-03-05 branch. A quick summary:

Jan Kundrát (7):
      Revert "spi: orion: Prepare space for per-child options"
      spi: orion: Respect per-transfer bits_per_word settings
      tty: max310x: Fail probe when external clock crystal is not stable
      leds: tlc591xx: SW reset during initialization
      pinctrl: mcp23s08: work around GPIO line naming
      hwmon: (max31790) Add OF device ID table
      igb: unbreak I2C bit-banging on i350

Václav Kubernát (7):
      hwmon: (max31790) Rework to use regmap
      hwmon: (max31790) Fix and split pwm*_enable
      hwmon: (max31790) Allow setting pulses
      hwmon: (max31790) Show 0 RPM/fault when input disabled
      hwmon: (max31790) Refactor HWMON_CHANNEL_INFO
      hwmon: (max31790) Allow setting fan*_div
      hwmon: (max31790) Update documentation

Change-Id: Ib61986bf64f0927e332fc31ba95b7340a79e1e09