Use rousette for RESTCONFish info retrieval
In the past we used a hacked-together set of shell scripts which invoked
`sysrepocfg` and some custom sysrepo wrappers in order to quickly get
JSON bits from the operational datastore. Switch from that to a little
more mature implementation of an HTTP server (which is still not really
RESTCONF).

There've been some changes in module revisions, so bump the YANG
revision dates as well. While doing that I discovered that the YANG
module installation is too complex; since 34d2f48 we do not preserve the
*.yang files anymore and we always start from an empty state. And while
I'm at this, always install the ietf-hardware module, that wouldn't hurt
on any box.

Change-Id: Ida87c3869bc538fe6c5fb597cf8624f2467ac0fe
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/4017
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/cla-sysrepo/+/3993
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/gammarus/+/4007
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/rousette/+/4013