Fix image builds for the VFAT USB flash
It's hard to fit 360+MB worth of data into a 350MB disk image.

The math is actually not as simple as taking a look at the rootfs sizes
or the on-disk consumption of the target/ directory because what ends up
on the USB disk image is the compressed CPIO root and a squashfs image.
Anyway, allow some breathing space here.

Change-Id: Ia54a9ae640b4f80a6311b9c234e131dfdb8d31e4