Make the Buildroot configuration public
It's a bit annoying that this repo references some other content
that is still private via git submodules, but at least this will
make it easier for us to release the full Buildroot configuration
and comply with the GPLv2. In fact, we're probably going an extra
mile here, but hey, it's handy and convenient, and there are no
NDAs covering this particular product.

Change-Id: I35be3565212c6e9cdb0d480d56d800d978e979a8