Update buildroot gdb
GDB 8.2 has a bug where it sometimes goes to background because of a
signal. This has happened to me when trying debug cla-sysrepo using
catchpoints. GDB 9.2 doesn't have this bug.

GDB 9.2 IS available in our version of buildroot, however, I wasn't able to
specify it. (I tried setting BR2_GDB_VERSION=9.2 to out defconfig but it
didn't do anything).

There is a buildroot patch that sets the default version to 9.2, so I
just applied that patch. The current buildroot tree can be found at
cesnet/2021-05-28-gdb9.

Bug: https://sourceware.org/legacy-ml/gdb-patches/2019-05/msg00361.html
Change-Id: Icf0278694f703a6bc59589de240949eb7d1976cd