Inline amp: fix duplicate netif names
The first production HW which has arrived uses a different brand of PCIe
NICs with two SFP ports. Our wildcard was too generic and matched both
interfaces, which is not what we want here. Adjust so that only the left
one gets matched.

While we're at this, also assign a different name to the last remaining
interface.

Change-Id: Ic6c20b9ab932221a63d837bc68e6e0ee8479f7ab