Make git repo detection cleverer
Currently, the script just checks whether we are inside a git
repository. However, in my setup, my home directory is a git repository
and that means that I have to run the script in an entirely different
place which is very annoying.

The new behavior is to check whether we are inside the same repository
as the dev-setup-git script.

Change-Id: I9072ad7a96f3521b470fd400d867c733a143b0d2