CI: also discard prebuilt C++ bindings for the NETCONF stack
When a job gets queued which updates, say, `cla-sysrepo` or
`netconf-cli`, a test build of the ARM image is queued. If the list of
changes that are being tested also includes an update to
`CzechLight/dependencies`, the intention was that "everything" which was
built via that update was also getting refreshed. That was, however,
actually not the case and the C++ bindings were reused from the prebuilt
artifact.

The same applies to other C++ libraries which are pinned via git
submodules. That looks like a reasonable compromise between the CI
latency and correctness.

Fixes: 8cd6156 Migrate to libyang2
Change-Id: I1a2ff559c600803d7ca464ced9f9dad20122d905