czechlight-cfg-fs: migration scripts and tests
Sometimes we need to make alterations to the startup datastore (for
instance when new feature to the YANG model is added).
This commit introduces a migration script that will be run (it is not
yet) after the machine boots. It versions current revision of the
startup datastore and when the backup of the startup DS found at /cfg
was of lower revision number, we apply these migrations scripts.

This commit introduces first migration. It merges initial data into
startup datastore.
These data are not supposed to be loaded at every boot. If the user
decides to drop them, then they are gone. Original behaviour restored
them at every boot.

We also introduce tests for these migrations scripts. TODO.

Change-Id: I96191630fd6b3a1ffbc5ba9f62c4a6c14ab5ed92