czechlight-cfg-fs: migration scripts and tests
Sometimes we need to make alterations to the startup datastore (for
instance when new feature to the YANG model is added).
This commit introduces a migration script that will be run (it is not
yet) after the machine boots. It versions current revision of the
startup datastore and when the backup of the startup DS found at /cfg
was of lower revision number, we apply these migrations scripts.

This commit introduces first migration. It merges initial data into
startup datastore.
These data are not supposed to be loaded at every boot. If the user
decides to drop them, then they are gone. Original behaviour restored
them at every boot.

We also introduce tests for these migrations scripts. They are located
under tests/czechlight-cfg-fs directory and there is a pytest runner.
Each test case is a single directory containing current startup DS
content (as JSON file), expected content after running migrations,
and three service files: One for current startup DS version (so we know
which migrations to run), one for cmdline (so both scripts
czechlight-install-yang and czechlight-migrate can distinguish between
czechlight box types in code), and xpath file used only if we want to
check for a part of the sysrepo DS content.

The tests simulate the workflow that is done when the system boots.
I.e., they install YANG models, restore startup DS backup into sysrepo
and then they run the migrations if necessary.

Change-Id: I96191630fd6b3a1ffbc5ba9f62c4a6c14ab5ed92