migration: add default configuration for ietf-interfaces
The current state of startup datastore for ietf-interfaces model did not
correspond to the real state of the box (its systemd-networkd
configuration). The datastore was empty, but its configuration was, by
default, set to "bridge all links" and set DHCP for both IPv4 and IPv6
for br0 (the bridge).

This migration should fix that. It overwrites the current networking
content in startup datastore with the values of the default
configuration described in the previous paragraph.

Change-Id: I0b809a39044bf6bcba4135a3456c684d94ef8b86