clearfog: update kernel to 6.2.9 + fixes
On one hand, this unbreaks the PCIe bus on the inline amplifiers (and
therefore the extra network card). On the other hand, there's a new
error from I2C:

 i2c i2c-2: readbytes: invalid block length (255)

This one might be harmless; at least I think it's a complain from a
piece of code that I touched the last time (!), and that was only to
make an error message grepable. Also, the buffer has been "always" set
to 32 bytes. Anyway, the initialization of the fsp-3y pmbus driver
passes, and I can read some data, which is good.

The other nasty part was a regression in drivers/tty/serial/max310x.c
which introduced random IO corruption when using batched IO (i.e., all
the time) -- hence a new patch.

I also decided to drop an unrelated spi_orion cleanup which isn't that
urgent -- so there's no change of the number in kernel's
CONFIG_LOCALVERSION.

Change-Id: I56470386d2c0abee508c883f15ea1986042349b5
Fixes: a12acd9 Update kernel