also restart velia-health when remounting sysrepo SHM
Since that daemon is *also* using sysrepo, it needs to be nuked properly
before we can remount the SHM tmpfs at /run/sysrepo. This was not a
problem in my testing because that daemon has died properly, but anyway,
it was missing.

The LEDs should work properly, and since sysrepo/NETCONF/RESTCONF are
our only other APIs for remote access (and they obviously need SHM to be
healthy), there's nothing else that we can do besides just killing
everything and starting from scratch, anyway.

Change-Id: I9a41018d14d7919fd5ebdcfe7bc8c0dbaecc8186