czechlight-cfg-fs: replace NACM restore service with a migration
Currently we reinitialize ietf-netconf-acm module after each boot by
importing a set of rules. This means that after each boot the NACM
configuration was overwritten with "our" configuration.
This is not what we want. Admin of our box can set his own NACM rules.
It makes sense to create a migration that sets the default NACM rules by
us and other people can modify them.

We scanned the git and gerrit history for the reason why NACM rules
were not moved to one of the first migrations but we found nothing on
that. So this is probably just one of the skeletons in the cupboard that
is being removed with this commit.

Change-Id: I084d6eb81fc340c09abc2edf2995ff40b219e807