Upgrade kernel to 6.5.9
Also available on Gerrit as cesnet/2023-10-26.

This requires patching buildroot due to upstream change in
libtraceevents, that's buildroot commit
b4ab45a5c1db05f84efb5f3b6909982263afb1d4.

Since kernel 6.4, the GPIO LED trigger is makred as "broken". I've
reported our use case back to the patch author (and gpio co-maintainer).

As of kernel 6.5, the DTS files got moved a little bit and are now
located below a vendor prefix in a subdirectory. Because we will be
jumping around multiple kernel versions during debugging of boot
failures, change this based on the actual version that's being compiled.
The kernel patch is of course still version-dependant, but that's not a
big problem.

Change-Id: Ia8fb582a759f16aa06f8d6484d09516363d1de8c
Bug: https://lore.kernel.org/all/20230314210059.419159-1-linus.walleij@linaro.org/