docs: use parallel build by default
...and tweak up the expected build time. The idea here is to give rough
expectations, and to gently guide the user towards enabling "some"
parallelization. Discussion of what exactly the `-j` parameter is doing
in my opinion doesn't belong here, so it's just a nudge to the user
which says "yup, it can do stuff in parallel", which was historically
not the case with upstream Buildroot. Hopefully, 8 job slots should be
"enough" for a substantial speedup, and "not too much" so as not to kill
any reasonable dev-sized machine (famous last words right there).

Change-Id: If88bc118af630bdcc39bea3d77a06c05a3c67a9a
Suggested-by: Michal Hažlinský <hazlinsky@cesnet.cz>