Re-enable the UTF-8 locale
On the fresh build straight from the CI, the `rousette` service was
crashing with an std::runtime_error:

 Could not set locale C.UTF-8

I think that this is caused by Buildroot purging the "unwated" locales
by default, and `systemd` enabling this locale as a default setting
based on the locale's availability on the build host. Now, while we have
glibc 2.35+ on the target system already, and on a build host as well,
so systemd enables this thing, the problem is that the final OS image
won't have this locale available, and therefore our C++ code will die
with an unhappy exception.

Fix this by making sure that the "reasonable" locales are available. I'm
tempted to disable all this locale purging altogether, but I feel
advantageous and want to waste hours of debugging time in future on
further locales, so I'm keeping the purging in as-is. Is that job
security, masochism, or what? Oh well, it's hot in this lab.

Fixes: 14498b76ca826437828c3d75afe205068e4736a6 (Remove systemd locale patch from our buildroot)
Change-Id: I0a4e72f6dfb4ab9523ed92bd9d9210cf09e6fee2